Use new SideNavigation for add patient
