+++ apps/modernization-ui/src/apps/page-builder/components/Navigation/PageBuilderSideNav.scss
@@ -1,0 +1,6 @@
+.side-nav {
+    display: flex;
+    border-radius: 5px;
+    border: 1px solid var(--base-lighter, #e6e6e6);
+    min-width: 245px;
+}
