+++ apps/modernization-ui/src/components/SideNavigation/NavigationEntry.tsx
@@ -1,0 +1,10 @@
+import { Link } from 'react-router-dom';
+
+type NavigationEntryProps = {
+    label: string;
+    href: string;
+    useNav?: boolean;
+};
+export const NavigationEntry = ({ href, label, useNav = false }: NavigationEntryProps) => {
+    return <>{useNav ? <Link to={href}>{label}</Link> : <a href={href}>{label}</a>}</>;
+};
