+++ apps/modernization-ui/src/pages/addPatient/components/DataEntrySideNav/DataEntrySideNav.scss
@@ -1,0 +1,3 @@
+.data-entry-side-nav {
+    border-right: 1px solid #dfe1e2;
+}
