--- apps/modernization-ui/src/pages/addPatient/components/LeftBar/MenuItem/MenuItem.scss
@@ -1,3 +1,0 @@
-.link {
-    color: black !important;
-}
