--- apps/modernization-ui/src/pages/addPatient/components/LeftBar/MenuItem/index.ts
@@ -1,1 +1,0 @@
-export * from './MenuItem';
